Changed inter-window communication to use localStorage events
window.opener was being cleared after visiting a different host
in Chrome 43 and Firefox 38. This change uses local storage
events + unique popup ids to communicate between windows. Here’s
an overview of how it works.

1. When popup.open is called, it generates a new popup id (UUID),
   and sets the popup window’s name to a serialized version of
   the popup id
2. When RedirectHandler#handle is called in a window with a valid
   serialized popup id as its name, it stores the serialized
   popup id in local storage, along with the url of the popup
   (which includes the auth code from the oauth provider)
3. When the new key/value is set in local storage, the popup
   service is notified of the StorageEvent. If StorageEvent#key
   contains the popup id we’re looking for, it extracts the
   auth code from the url, resolves the promise, and removes the
   key from local storage.

Other notes

* Don’t need to worry about opening popup with _blank name
anymore, because that was just a work around due to Chrome
on iOS returning null for window.opener (which is no longer used)