Remove use of window.name as a channel for localStorage messages
window.name was used to decide what item key would be used for a popup's
message to the parent. This is nor reliable, as some browsers leak the
fact that Facebook (and other popups) may set a window name that
destroys the one set by Torii.

Instead, this patch uses a global key set by torii, identifying a
"current channel" that expects to receive the message. This is not
perfect, as a window opening a torii request after an different window
may receve the completed auth from the window of the first request.
However in a practical sense it means the last window starting auth will
receive the messages from auth popups, and old windows that may have
started auth will receive none.