facebook-connect - Add support for auth_type
Adds the ability to set the `auth_type` when connecting to FB. Currently
FB only supports `rerequest` which allows for requesting declined 
permissions.

This is not suitable for global configuration because you may not always
want to rerequest permissions but do it based on other conditions. As
such its passed through to the open function in options. It defaults
to not being set aka the standard facebook-connect auth_type.