Add CSRF protection for OAuth2 providers via state
The OAuth2 code provider (and all subclasses) now default to a
generated random "state" URL parameter. This is transmitted to the
server, and the provider validates that the same state parameter is
received in the response.

This protects against Cross-Site Request Forgery. For more details on
this part of the OAuth2 spec, see here RFC 6749 section 10.12:

http://tools.ietf.org/html/rfc6749#section-10.12