Fix OAuth2 state validation error
Need to add the "state" parameter to responseParams in order for the
popup code to resolve it.

Since it is optional and must be implemented by the oauth2-code
subclasses, we don't now don't validate the state parameter unless it is
present in the list of response Params.

Update the mock popup to reflect this behaviour, only sending back a
state value if it is in the passed set of keys.